Separate Filtermenu state from assign section state
